refactor: lib 폴더 적용
