+++ src/pages/container/ContainerExample.tsx
@@ -1,10 +1,10 @@
-import Container from "../../layouts/Container/Container.tsx";
+import {Container} from "../../../lib";
 
 function ContainerExample() {
   return (
     <>
       <Container>
-        <div style={{ background: "lightblue" }}>
+        <div style={{background: "lightblue"}}>
           <h1>Your content goes here</h1>
           <p>
             This is a container with dynamic maxWidth and a light blue
@@ -13,7 +13,7 @@ function ContainerExample() {
         </div>
       </Container>
       <Container maxWidth="sm">
-        <div style={{ background: "lightblue" }}>
+        <div style={{background: "lightblue"}}>
           <h1>Your content goes here</h1>
           <p>
             This is a container with dynamic maxWidth and a light blue
@@ -22,7 +22,7 @@ function ContainerExample() {
         </div>
       </Container>
       <Container maxWidth="md">
-        <div style={{ background: "lightblue" }}>
+        <div style={{background: "lightblue"}}>
           <h1>Your content goes here</h1>
           <p>
             This is a container with dynamic maxWidth and a light blue
@@ -31,7 +31,7 @@ function ContainerExample() {
         </div>
       </Container>
       <Container maxWidth="lg">
-        <div style={{ background: "lightblue" }}>
+        <div style={{background: "lightblue"}}>
           <h1>Your content goes here</h1>
           <p>
             This is a container with dynamic maxWidth and a light blue
@@ -40,7 +40,7 @@ function ContainerExample() {
         </div>
       </Container>
       <Container maxWidth="xl">
-        <div style={{ background: "lightblue" }}>
+        <div style={{background: "lightblue"}}>
           <h1>Your content goes here</h1>
           <p>
             This is a container with dynamic maxWidth and a light blue
@@ -49,7 +49,7 @@ function ContainerExample() {
         </div>
       </Container>
       <Container maxWidth="xxl">
-        <div style={{ background: "lightblue" }}>
+        <div style={{background: "lightblue"}}>
           <h1>Your content goes here</h1>
           <p>
             This is a container with dynamic maxWidth and a light blue
@@ -58,7 +58,7 @@ function ContainerExample() {
         </div>
       </Container>
       <Container minWidth={600} maxWidth={960}>
-        <div style={{ background: "lightblue" }}>
+        <div style={{background: "lightblue"}}>
           <h1>Your content goes here</h1>
           <p>
             This is a container with dynamic maxWidth and a light blue
